# Improve efficiency of arranging nodes